Split bootloader implementation into multiple files
